+++ core/main_method/storage.py
@@ -24,15 +24,13 @@ from core.libs import parse_params
 
 
 def _clearStorage(self):
-    """After saved storage,
-    clean all the item of two table widgets.
-    """
+    """After saved storage, clean all the item of two table widgets."""
     self.EntitiesPoint.clear()
     self.EntitiesLink.clear()
     self.InputsWidget.variableExcluding()
 
 
-def _addStorage(self, name, expr):
+def _addStorage(self, name: str, expr: str):
     """Add storage data function."""
     self.CommandStack.beginMacro("Add {{Mechanism: {}}}".format(name))
     self.CommandStack.push(AddStorage(
@@ -152,6 +150,14 @@ def on_mechanism_storage_restore_clicked(self, item: QListWidgetItem = None):
         self.mechanism_storage_name_tag
     ))
     self.CommandStack.endMacro()
+
+
+def getStorage(self) -> Tuple[Tuple[str, str]]:
+    """Get storage data."""
+    return tuple((
+        self.mechanism_storage.item(row).text(),
+        self.mechanism_storage.item(row).expr
+    ) for row in range(self.mechanism_storage.count()))
 
 
 def loadStorage(self, exprs: Tuple[Tuple[str, str]]):